refactor(auth-visit): Split master-data and visitSchedule routes out of controllers
auth-service's master-data.controller.ts and visit-form-service's
visitSchedule.controller.ts move route/middleware declarations and
response Zod schemas into their own *.routes.ts files; controllers keep
only request-handling logic via createXController(service) factories.
module.ts files now create the shared DocumentedRouter and register
routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after for both services, lint clean, all 239 (auth) and 266
(visit-form) existing tests pass, and live smoke checks confirming
middleware order (auth before oversized-batch check) is preserved.
