+++ apps/auth-service/src/master-data/master-data.module.ts
@@ -1,11 +1,12 @@
-import type { DocumentedRouter, TokenSigner } from '@armman/service-commons';
+import type { TokenSigner } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { GeographyRepository } from '../geography/geography.repository';
 import { ProjectRepository } from '../projects/project.repository';
 import { MasterDataService } from './master-data.service';
-import { createMasterDataRouter } from './master-data.controller';
+import { registerMasterDataRoutes } from './master-data.routes';
 
-/** Composition root for the master-data delta feature: wires repositories → service → router. */
+/** Composition root for the master-data delta feature: wires repositories → service → routes. */
 export function createMasterDataModule(
   prisma: PrismaService,
   signer: TokenSigner,
@@ -13,5 +14,7 @@ export function createMasterDataModule(
   const geographyRepository = new GeographyRepository(prisma);
   const projectRepository = new ProjectRepository(prisma);
   const service = new MasterDataService(geographyRepository, projectRepository);
-  return createMasterDataRouter(service, signer);
+  const doc = createDocumentedRouter();
+  registerMasterDataRoutes(doc, service, signer);
+  return doc;
 }
